+++ Editor/LauncherAttribute.cs
@@ -1,9 +1,0 @@
-﻿namespace EasyEditor
-{
-    internal class LauncherAttribute : System.Attribute
-    {
-        public string[] InstallationNames { get; }
-        public LauncherAttribute(params string[] installationNames) { InstallationNames = installationNames; }
-    }
-}
-