+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-09-20
+### Added
+- Per launcher settings.
+- Tooltips on settigns.
+
+### Changed
+- Robust code editor installation matching.
+
 ## [0.2.0] - 2020-09-16
 ### Added
 - Ability to match a proper C# compiler version: 7.3 for 2020.1 and below, 8.0.